requirements.txt
imdbpy
numpy
pymongo
tensorflow
tqdm
wandb